+++ hub-app/src/main/java/io/fulflix/hub/hubroute/api/dto/HubRouteCreateDto.java
@@ -1,4 +1,4 @@
-package io.fulflix.hub.application.dto;
+package io.fulflix.hub.hubroute.api.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
