+++ hub-app/src/main/java/io/fulflix/hub/hubroute/api/dto/HubRouteResponseDto.java
@@ -1,4 +1,4 @@
-package io.fulflix.hub.application.dto;
+package io.fulflix.hub.hubroute.api.dto;
 
 import lombok.*;
