+++ hub-app/src/main/java/io/fulflix/hub/hubroute/domain/HubRoute.java
@@ -1,5 +1,6 @@
-package io.fulflix.hub.domain;
+package io.fulflix.hub.hubroute.domain;
 
+import io.fulflix.hub.hub.domain.Hub;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
